Renamed JSONReader to ModJSONParser for clarity; packaged JSON-related files together
